+++ scripts/characters/State.gd
@@ -1,7 +1,6 @@
 class_name SMState
 extends Node
 
-
 # SAMPLE STATE:
 # copy and paste this code into a script on a new node as a child of this node
 # name the node what you want the state to be changed
@@ -12,14 +11,3 @@ extends Node
 #@export var max_speed: int
 #@export var acceleration: int
 #@onready var sm = get_parent()
-
-
-#func update_state(delta):
-	#print("I run every frame when this is the current frame")
-	#pass
-#func enter_state():
-	#print ("I run once when the state machine changes to this state")
-	#pass
-#func exit_state():
-	#print(" I run once when this is the current state and change state is called")
-	#pass