+++ scripts/for_scenes/HUD.gd
@@ -11,6 +11,7 @@ var test = 0
 # Called when the node enters the scene tree for the first time.
 func _ready() -> void:
 	SignalBus.player_stats_updated.connect(update_player_info)
+	SignalBus.player_health_updated.connect(update_health)
 
 # Called every frame. 'delta' is the elapsed time since the previous frame.
 func _process(delta: float) -> void:
@@ -26,4 +27,8 @@ func toggle_upgrade_menu():
 func update_player_info():
 	#Updates health bar when player upgrades health
 	health_bar.max_value = player.health_manager.max_health
+	update_health()
+	
+func update_health():
+	health_bar.value = player.health_manager.health
 	health_label.text = "%d/%d" % [player.health_manager.health, player.health_manager.max_health]