+++ scripts/globals/globals.gd
@@ -1,6 +1,5 @@
 extends Node
 
-
 var player: CharacterBody2D = null
 
 var dir_dict = {