+++ scripts/globals/upgrade_manager.gd
@@ -1,18 +1,32 @@
 extends Node 
 
 #upgradable stats
-var health_lvl = 1
-var dmg_lvl = 1
-var speed_lvl = 1
+# TODO
+# delink level from stat increase, convert to indices of an upgrades array (source of truth)
+# hp and dmg are flat increases, starting at 0
+var health_lvl: int = 0
+var dmg_lvl: int = 0
+# speed is a mult increase, should start at 1
+var speed_lvl: int = 1
 
 #amount to increase upgrades by
 @export_category("Upgrade Growth")
 @export var health_up = 1
 @export var dmg_up = 1
 @export var speed_up = 1.2
+@export var upgrade_cost = 1
+@export var currency_name: String = "money"
+
+#Inventory ref from player
+@onready var inventory: Inventory = get_tree().get_first_node_in_group("Inventory")
 
 func _ready():
 	SignalBus.upgrade_stat_button_pressed.connect(_on_stat_upgraded)
+	
+func check_can_upgrade(upgrade_lvl: int) -> bool:
+	if upgrade_lvl+1 * upgrade_cost <= inventory.check_item(currency_name):
+		return true
+	return false
 
 # ALL STATS SHOULD HAVE A METHOD ACCORDING TO THIS CONVENTION
 # func upgrade_{stat_name}():
@@ -48,6 +62,7 @@ func get_speed_lvl():
 	return speed_lvl
 
 func _on_stat_upgraded(stat_name: String):
-	Callable(self, "upgrade_%s" %stat_name).call()
+	if check_can_upgrade(Callable(self, "get_%s_lvl" %stat_name).call()):
+		Callable(self, "upgrade_%s" %stat_name).call()
 	#notify player to update self
 	SignalBus.player_stats_updated.emit()