+++ scripts/pit.gd
@@ -1,0 +1,6 @@
+extends Area2D
+
+func _on_area_entered(area):
+	print("player entered")
+	
+	pass # Replace with function body.