+++ scripts/tile.gd
@@ -1,0 +1,4 @@
+extends Node
+class_name Tile
+
+var type = null