+++ CHANGELOG.md
@@ -39,6 +39,14 @@
 
 - Added player upgrades
 
+# version 0.17
+
+- Level Spawn Package created
+- Enemy Pooling Implemented
+- Hitbox/ Hurtbox balancing revision
+- Export settings configured for playtesting
+- Mobility Manager fixed to properly use the max_speed
+
 # version 0.18
 
 - Added background to title scene